No ACIA/IKBD SSE mods if option '6301' not checked
This is because timings were wrong but the system was somehow adjusted
to them. We can fix some parts of fake ikbd emulation but then some
programs will be reported as not working anymore.

So now, when 6301 option isn't checked, you should get exactly the
same keyboard emulation as in v3.2.
When this option is checked, ACIA emulation is changed, for the
IKBD as well as for MIDI.

This also has the advantage of streamlining the source.
Old code parts have been removed for readability (adding yet more 
switches to this mess is almost impossible).


git-svn-id: svn://svn.code.sf.net/p/steemsse/code/trunk/steemsse@188 dd3fbd3b-ef5d-4b78-98e8-da144fa02e5b